+++ src/Models/PatchProductBinding.go
@@ -1,9 +1,9 @@
 package Models
 
 type PatchProductBinding struct {
-	Name     *string `json:"Name"`
-	Category *string `json:"Category"`
-	Price    *uint   `json:"Price"`
-	Quantity *uint   `json:"Quantity"`
-	ImageURL *string `json:"ImageURL"`
+	Name     *string  `json:"Name"`
+	Category *string  `json:"Category"`
+	Price    *float64 `json:"Price"`
+	Quantity *uint    `json:"Quantity"`
+	ImageURL *string  `json:"ImageURL"`
 }
